RUB-270 reduce go node storage helper complexity
